api-deprecated/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-core==1.32.0
azure-cosmos==4.9.0
azure-functions==1.21.3
azure-identity==1.20.0
azure-storage-blob==12.24.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
fastapi==0.115.8
h11==0.14.0
idna==3.10
isodate==0.7.2
msal==1.31.1
msal-extensions==1.2.0
portalocker==2.10.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
